team1/requirements.txt
djangorestframework
Pillow